+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## master
+
+- removed auto positioning (did not like implementation and not really sure if it is needed)
+- tooltip content does not have to be with-in the main tooltip element
+- changed how to define tooltip positioning
+- a tooltip can now have both a regular hover tooltip content and a click sticky tooltip content
+
 ## 0.6.0
 
 - added .editorconfig file